# bank-product-service/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Swagger Products Bank - OpenAPI 3.0
  description: |-
    This is a sample Pet Store Server based on the OpenAPI 3.0 specification.  You can find out more about
    Swagger at [https://swagger.io](https://swagger.io). In the third iteration of the pet store, we've switched to the design first approach!
    You can now help us improve the API whether it's by making changes to the definition itself or to the code.
    That way, with time, we can improve the API in general, and expose some of the new features in OAS3.
  termsOfService: http://swagger.io/terms/
  contact:
    email: apiteam@swagger.io
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.11
...
servers:
  - url: http://localhost:8001/api/v1
...
paths:
...
  /products/{productId}:
    delete:
      tags:
        - product
      summary: Delete a product
      description: delete a product
      operationId: deleteBy
      parameters:
        - name: productId
          in: path
          description: product id to delete
          required: true
          schema:
            type: integer
      responses:
        '400':
          description: Invalid product value
...